src/main/java/com/github/mikesafonov/specification/builder/starter/predicates/NotPredicateBuilder.java
package com.github.mikesafonov.specification.builder.starter.predicates;

import lombok.RequiredArgsConstructor;

import javax.persistence.criteria.CriteriaBuilder;
import javax.persistence.criteria.CriteriaQuery;
import javax.persistence.criteria.Predicate;
import javax.persistence.criteria.Root;

@RequiredArgsConstructor
public class NotPredicateBuilder implements PredicateBuilder {

    private final PredicateBuilder predicateBuilder;

    @Override
    public Predicate build(Root<?> root, CriteriaQuery<?> q, CriteriaBuilder cb) {
        return cb.not(predicateBuilder.build(root, q, cb));
    }
}
